Remove horizontal and vertical dragging detection
Dragging orientation detection conflicts with other gestures. Therefore
guesture detection of any kind is outside the scope of pointer-event.lua

It's better to implement that in examples, where people will use to
start off with anyway.

This has the advantage that dragging start doesn't carry doesn't add
anything to the command anymore, making it usable outside of script
messages.