Testing/log-replay things
The mocking/testing stuff has been developing into a log replaying
direction.

Unfortunately the requirements for mocking and log replay are
different enough that those will need to be split into separate
projects.

The upside of that is that initial experimentation has shown that
log replay should be fairly straight forward to do, enabling debugging
the running script as if it was running in the mpv from the user that
provided the log.
The basic idea is to overwrite all builtin function with side
effects with my own wrappers that log the necessary data. During
replay that data is then used to call the correct callbacks at the right
time, provide the same property values like they were while running in
mpv etc.